Se arreglan los problemas con el alta de eventos, se agrega la posibilidad de fechas futuras en el alta de eventos para las fechas posibles de germinacion y estratificacion
