+++ app/src/main/java/ar/edu/ort/bmon/rootsapp/ui/event/CreateEventFragment.java
@@ -305,7 +305,7 @@ public class CreateEventFragment extends Fragment {
     }
 
     private void showDatePickerDialog(FragmentManager fragmentManager) {
-        DatePickerFragment newFragment = DatePickerFragment.newInstance(new DatePickerDialog.OnDateSetListener() {
+        DatePickerFragment newFragment = DatePickerFragment.newInstance(true, new DatePickerDialog.OnDateSetListener() {
             @Override
             public void onDateSet(DatePicker datePicker, int year, int month, int day) {
                 // +1 because January is zero
