+++ app/src/main/java/ar/edu/ort/bmon/rootsapp/ui/event/EventDetailFragment.java
@@ -481,7 +481,7 @@ public class EventDetailFragment extends DialogFragment {
     }
 
     private void showDatePickerDialog(FragmentManager fragmentManager, final EditText field) {
-        DatePickerFragment newFragment = DatePickerFragment.newInstance(new DatePickerDialog.OnDateSetListener() {
+        DatePickerFragment newFragment = DatePickerFragment.newInstance(true, new DatePickerDialog.OnDateSetListener() {
             @Override
             public void onDateSet(DatePicker datePicker, int year, int month, int day) {
                 // +1 because January is zero
