+++ app/src/main/java/ar/edu/ort/bmon/rootsapp/ui/plant/CreatePlantFragment.java
@@ -526,7 +526,7 @@ public class CreatePlantFragment extends Fragment {
     }
 
     private void showDatePickerDialog(FragmentManager fragmentManager) {
-        DatePickerFragment newFragment = DatePickerFragment.newInstance(new DatePickerDialog.OnDateSetListener() {
+        DatePickerFragment newFragment = DatePickerFragment.newInstance(false, new DatePickerDialog.OnDateSetListener() {
             @Override
             public void onDateSet(DatePicker datePicker, int year, int month, int day) {
                 // +1 because January is zero
