+++ app/src/main/res/layout/fragment_germination_options.xml
@@ -33,7 +33,7 @@
     </androidx.cardview.widget.CardView>
 
     <EditText
-        android:id="@+id/editTextFechaEstratificacion"
+        android:id="@+id/editTextFechaEstrata"
         android:layout_width="190dp"
         android:layout_height="50dp"
         android:layout_marginTop="8dp"
